schemaprinter-comparators Minor refactoring from PR review
